+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Fixed
+
+- New Topic validation errors are now properly aligned with the form
+
 ## [0.6] - 2021-09-16
 
 ### Added
@@ -170,6 +176,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - User cache for reduced DB queries
 - Page title updates to reflect current view
 
+[Unreleased]: https://github.com/APB9785/phxBB/compare/0.6...HEAD
 [0.6]: https://github.com/APB9785/phxBB/compare/0.5.2...0.6
 [0.5.2]: https://github.com/APB9785/phxBB/compare/0.5.1...0.5.2
 [0.5.1]: https://github.com/APB9785/phxBB/compare/0.5...0.5.1